+++ internal/logapi/repository.go
@@ -340,6 +340,7 @@ func optionalInt64(value *int64) string {
 }
 
 type commonLogRecord struct {
+	charityModel                                                *string
 	rejectionStage, rejectionReason, requestMethod, requestPath sql.NullString
 	rowID                                                       int64
 	id                                                          string